feat: implement local storage utilities for MapCanvas state persistence
